Remove tabela device e ajusta as demais classes

+++ src/main/resources/db/migration/V202310250013__alter_users_table.sql
@@ -1,4 +1,4 @@
 ALTER TABLE users
 ADD COLUMN IF NOT EXISTS device_token varchar(255),
-ADD COLUMN IF NOT EXISTS last_latitude decimal(10, 8),
-ADD COLUMN IF NOT EXISTS last_longitude decimal(11, 8);
+ADD COLUMN IF NOT EXISTS last_latitude float(53),
+ADD COLUMN IF NOT EXISTS last_longitude float(53);
